requirements.txt
# 请使用 pip 安装；勿将 conda/pip freeze 中带 file:// 本地路径的条目提交到此文件。
# 若需锁定全量依赖，可在干净虚拟环境中 pip freeze > requirements.lock.txt

annotated-types==0.7.0
anthropic==0.34.2
anyio==4.12.1
asgiref==3.11.0
astor==0.8.1
brotlicffi>=1.1.0.0
certifi>=2024.8.30
cffi>=1.17.0
charset-normalizer>=3.3.0
click==8.3.1
distro==1.9.0
Django==4.2.9
django-admin-black==1.0.19
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
django-unfold==0.81.0
djangorestframework==3.15.2
exceptiongroup==1.3.1
filelock>=3.16.0
graphviz==0.20.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna>=3.10
Jinja2>=3.1.4
jiter==0.12.0
MarkupSafe>=2.1.5
mpmath>=1.3.0
networkx>=3.2
numpy>=2.0,<2.3
packaging>=24.0
pandas==2.2.3
pillow==10.2.0
psycopg2-binary>=2.9.9
pycparser>=2.22
pydantic==2.12.5
pydantic_core==2.41.5
PySocks>=1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML>=6.0.1
requests==2.32.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
str2bool==1.1
sympy>=1.13
tokenizers==0.22.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions>=4.12.0
tzdata==2025.3
urllib3>=2.2.0
whitenoise==6.7.0
